Changed table_id to use a size_t for the index argument, which is both semantically correct and fixes the conversion issue in fifo_thread.c

+++ src/util/common_table.c
@@ -201,7 +201,7 @@ TableResponseStatus table_id_count(TableType type, size_t *count)
   return OK;
 }
 
-TableResponseStatus table_id(TableType type, int index, charbuf *id)
+TableResponseStatus table_id(TableType type, size_t index, charbuf *id)
 {
   Table *table = get_table_by_type(type);
